update the script to process events from the dispatch stream
Previously, one script would connect to the dispatcher and dispatch events
round robin style over a zeromq socket to workers.

Occasionally for some reason the stream page would show something like only
half the event rate. I was unable to debug why this was happening, but I
suspect that it has something to do with the zeromq sockets (see
http://lucumr.pocoo.org/2012/6/26/disconnects-are-good-for-you/).

To fix this issue, there is now only one script which connects directly to the
dispatch stream and processes the events. To allow you to run multiple scripts
to process the data in parallel, the script now takes two arguments to tell it
the number of workers and the worker number for each script. Then, each worker
only processes events in which gtid % num_workers == worker number.