update cmos and base current scripts
This commit updates the CMOS rate and base current scripts to directly connect
to the data server instead of using a round robin routing scheme with zeromq.
The initial round robin scheme was designed so that the scripts would never lag
behind, i.e. even if you were polling the cmos rates for every crate as fast as
you could, the data on the monitoring site would be current since there were
always 8 workers parsing the data.

In practice, however, I'm not sure that it's so much better to use multiple
workers in the way that it was being used. The reason is that, at least in the
current scheme, where each worker parses data from all crates, it is necessary
to store the CMOS counts and timestamps also in redis (not just the rates), and
to fetch the latest counts and timestamps before calculating the rates. This
round trip latency is probably significant, and if you only have one worker you
can just store the counts locally, and so you never have to query the database
to find out the last cmos counts.

Note: The big reason for updating this is that the CMOS rate and base current
scripts seem to quit working randomly. I have not been able to debug what is
going on, but suspect it has to do with zeromq. See
http://lucumr.pocoo.org/2012/6/26/disconnects-are-good-for-you/.