update the time series steps on the stream and L2 page so that they are the same as the buckets in the database
Chris pointed out that he noticed an orphan on the L2 page appearing and
dissappearing on the L2 page when he would refresh the page. The reason for
this is that the steps in the drop down menu are not identical to the buckets
stored in the database. Therefore, when you request the time series with a step
size of 3 for example, but the database only has the data stored in 2 and 4
second buckets, it will return to you the data from the 2 second buckets but
occasionally skip a data point. This is currently noticeable for example when
you look at the PONG and SYNC signals on the 1 day timescale you see that there
appear to be occasional missing signals. However, this is just due to the
skipping of occasional bins since the step size for 1 day did not match up
perfectly with the bucket sizes in the database.

This update changes the step sizes to be the same as those stored in redis and
also adds two more bucket sizes of 2**17 and 2**18 seconds so that the data
will be stored in the redis database for ~20 years instead of just 4. I did
this since people are still analyzing SNO data and it's been ~20 years since
they started, so I thought it would be a good idea to have the SNO+ data also
stored for at least that long, although who knows if minard will still be
running in 20 years!

Thanks to Chris Connors for pointing out this bug!