Add a check that the event times of the
muons, atmospherics, and missed muons come within the time
range defined by the stop and start times of the run tables.
This is calculated on the nearline and the time_check table
is filled with that information. This also changes the event
times to be displayed in local time (EST) for direct comparison
to the runs page.