fix get_alarms() function
Previously if an alarm was still active it might not show up in the list of
alarms for runs since we looked for overlap between the run start and stop time
and the greatest of the cleared and acknowledge times. But, if an alarm had
been acknowledged but not cleared then we would only check for overlap with the
cleared time which would cause the alarm to not show up until it had been
cleared.

To fix this, we use the COALESCE postgres function to set the cleared and/or
acknowledged times to the current time if they are NULL. Also switch to using
the postgres OVERLAPS keyword to check if an alarm occurred during a given run.
Pretty sweet function!