fix a bug and indent noise page
This commit fixes a bug on the PMT noise page which was causing the dropdown
links in the top menu to not work. I eventually tracked it down to jquery being
imported in the script block even though it's already imported in the layout
file.

I also indented the PMT noise html page and made sure it had all the correct
rows and col-md-12's.